+++ EjemploGUI/src/ejemplogui/Profesor.java
@@ -11,7 +11,7 @@ package ejemplogui;
  */
 public class Profesor extends Persona{
 
-    public Profesor(String name, String lastName, int landLine, int movilPhone, String mail, String address) {
+    public Profesor(String name, String lastName, long landLine, long movilPhone, String mail, String address) {
         super(name, lastName, landLine, movilPhone, mail, address);
     }
